Buttons should be explicitly defined in mdx for now
We don't like every file having to define its own buttons and always
searching for the last link in a component (for example) is janky.

Providing buttonText and buttonHref to a component like CTA does is also
janky. We need one way of doing it. We should write them out in mdx
explicitly.

[[Link text]](href)

You can also do props. Some examples:

[[Sign up now]](/sign-up)
[[Sign up now|size=large,variant=outlined]](/sign-up)
[[See pricing]](#pricing)